OLS-1215: Refactored user feedback tests
