+++ tests/e2e/test_api.py
@@ -216,136 +216,6 @@ def test_forbidden_user():
 
 
 @pytest.mark.cluster()
-def test_feedback_can_post_with_wrong_token():
-    """Test posting feedback with improper auth. token."""
-    response = client.post(
-        "/v1/feedback",
-        json={
-            "conversation_id": CONVERSATION_ID,
-            "user_question": "what is OCP4?",
-            "llm_response": "Openshift 4 is ...",
-            "sentiment": 1,
-        },
-        timeout=BASIC_ENDPOINTS_TIMEOUT,
-        headers={"Authorization": "Bearer wrong-token"},
-    )
-    assert response.status_code == requests.codes.forbidden
-
-
-@pytest.mark.cluster()
-def test_feedback_storing_cluster():
-    """Test if the feedbacks are stored properly."""
-    feedbacks_path = OLS_USER_DATA_PATH + "/feedback"
-    pod_name = cluster_utils.get_pod_by_prefix()[0]
-
-    # disable collector script to avoid interference with the test
-    cluster_utils.create_file(pod_name, OLS_COLLECTOR_DISABLING_FILE, "")
-
-    # there are multiple tests running agains cluster, so transcripts
-    # can be already present - we need to ensure the storage is empty
-    # for this test
-    feedbacks = cluster_utils.list_path(pod_name, feedbacks_path)
-    if feedbacks:
-        cluster_utils.remove_dir(pod_name, feedbacks_path)
-        assert cluster_utils.list_path(pod_name, feedbacks_path) is None
-
-    response = client.post(
-        "/v1/feedback",
-        json={
-            "conversation_id": CONVERSATION_ID,
-            "user_question": "what is OCP4?",
-            "llm_response": "Openshift 4 is ...",
-            "sentiment": 1,
-        },
-        timeout=BASIC_ENDPOINTS_TIMEOUT,
-    )
-
-    assert response.status_code == requests.codes.ok
-
-    feedback_data = cluster_utils.get_single_existing_feedback(pod_name, feedbacks_path)
-
-    assert feedback_data["user_id"]  # we don't care about actual value
-    assert feedback_data["conversation_id"] == CONVERSATION_ID
-    assert feedback_data["user_question"] == "what is OCP4?"
-    assert feedback_data["llm_response"] == "Openshift 4 is ..."
-    assert feedback_data["sentiment"] == 1
-
-
-def test_feedback_missing_conversation_id():
-    """Test posting feedback with missing conversation ID."""
-    response = client.post(
-        "/v1/feedback",
-        json={
-            "user_question": "what is OCP4?",
-            "llm_response": "Openshift 4 is ...",
-            "sentiment": 1,
-        },
-        timeout=BASIC_ENDPOINTS_TIMEOUT,
-    )
-
-    response_utils.check_missing_field_response(response, "conversation_id")
-
-
-def test_feedback_missing_user_question():
-    """Test posting feedback with missing user question."""
-    response = client.post(
-        "/v1/feedback",
-        json={
-            "conversation_id": CONVERSATION_ID,
-            "llm_response": "Openshift 4 is ...",
-            "sentiment": 1,
-        },
-        timeout=BASIC_ENDPOINTS_TIMEOUT,
-    )
-
-    response_utils.check_missing_field_response(response, "user_question")
-
-
-def test_feedback_missing_llm_response():
-    """Test posting feedback with missing LLM response."""
-    response = client.post(
-        "/v1/feedback",
-        json={
-            "conversation_id": CONVERSATION_ID,
-            "user_question": "what is OCP4?",
-            "sentiment": 1,
-        },
-        timeout=BASIC_ENDPOINTS_TIMEOUT,
-    )
-
-    response_utils.check_missing_field_response(response, "llm_response")
-
-
-def test_feedback_improper_conversation_id():
-    """Test posting feedback with improper conversation ID."""
-    response = client.post(
-        "/v1/feedback",
-        json={
-            "conversation_id": "incorrect-conversation-id",
-            "user_question": "what is OCP4?",
-            "llm_response": "Openshift 4 is ...",
-            "sentiment": 1,
-        },
-        timeout=BASIC_ENDPOINTS_TIMEOUT,
-    )
-
-    # error should be detected on Pydantic level
-    assert response.status_code == requests.codes.unprocessable
-
-    # for incorrect conversation ID, the payload should be valid JSON
-    response_utils.check_content_type(response, "application/json")
-    json_response = response.json()
-
-    assert (
-        "detail" in json_response
-    ), "Improper response format: 'detail' node is missing"
-    assert (
-        json_response["detail"][0]["msg"]
-        == "Value error, Improper conversation ID incorrect-conversation-id"
-    )
-
-
-@pytest.mark.cluster()
 def test_transcripts_storing_cluster():
     """Test if the transcripts are stored properly."""
     transcripts_path = OLS_USER_DATA_PATH + "/transcripts"
